Fixed preloader bug
The particle and snowflake effect has been fixed in all the side pages.

The issue was that the preloader.css file had "display none" for the content that affected the particle effect, and the snowflake effect was preceeded by warning modal in script.js so that cause error in console.

Ive made a separate file for the preloader js and added the snowflake function in it so that all the pages have the preloader as well as the snowflake thing